Upload Function source into region storage service

+++ rust-libs/storage/src/lib.rs
@@ -1,7 +1,7 @@
 use anyhow::{bail, Error, Result};
 use async_trait::async_trait;
 use dyn_clonable::clonable;
-use mu_stack::StackID;
+use mu_stack::{StackID, StackOwner};
 use pin_project_lite::pin_project;
 use s3::{creds::Credentials, Bucket};
 use serde::Deserialize;
@@ -16,24 +16,39 @@ pub struct Object {
     pub size: u64,
 }
 
+#[derive(Clone, Copy, Hash, PartialEq, Eq)]
+pub enum Owner {
+    User(StackOwner),
+    Stack(StackID),
+}
+
+impl Owner {
+    fn path_prefix(&self) -> String {
+        match self {
+            Owner::User(pk) => format!("u!{pk}"),
+            Owner::Stack(id) => format!("s!{id}"),
+        }
+    }
+}
+
 #[async_trait]
 #[clonable]
 pub trait StorageClient: Send + Sync + Clone {
     async fn update_stack_storages(
         &self,
-        stack_id: StackID,
+        owner: Owner,
         storage_delete_pairs: Vec<(&str, DeleteStorage)>,
     ) -> Result<()>;
 
-    async fn storage_list(&self, stack_id: StackID) -> Result<Vec<String>>;
+    async fn storage_list(&self, owner: Owner) -> Result<Vec<String>>;
 
-    async fn contains_storage(&self, stack_id: StackID, storage_name: &str) -> Result<bool>;
+    async fn contains_storage(&self, owner: Owner, storage_name: &str) -> Result<bool>;
 
-    async fn remove_storage(&self, stack_id: StackID, storage_name: &str) -> Result<()>;
+    async fn remove_storage(&self, owner: Owner, storage_name: &str) -> Result<()>;
 
     async fn get(
         &self,
-        stack_id: StackID,
+        owner: Owner,
         storage_name: &str,
         key: &str,
         writer: &mut (dyn AsyncWrite + Send + Sync + Unpin),
@@ -41,20 +56,15 @@ pub trait StorageClient: Send + Sync + Clone {
 
     async fn put(
         &self,
-        stack_id: StackID,
+        owner: Owner,
         storage_name: &str,
         key: &str,
         reader: &mut (dyn AsyncRead + Send + Sync + Unpin),
     ) -> Result<()>;
 
-    async fn delete(&self, stack_id: StackID, storage_name: &str, key: &str) -> Result<()>;
+    async fn delete(&self, owner: Owner, storage_name: &str, key: &str) -> Result<()>;
 
-    async fn list(
-        &self,
-        stack_id: StackID,
-        storage_name: &str,
-        prefix: &str,
-    ) -> Result<Vec<Object>>;
+    async fn list(&self, owner: Owner, storage_name: &str, prefix: &str) -> Result<Vec<Object>>;
 }
 
 #[derive(Clone, Debug)]
@@ -120,8 +130,8 @@ impl StorageClientImpl {
         Ok(StorageClientImpl { bucket })
     }
 
-    fn create_path(stack_id: StackID, storage_name: &str, key: &str) -> String {
-        format!("{stack_id}/{storage_name}/{key}")
+    fn create_path(owner: Owner, storage_name: &str, key: &str) -> String {
+        format!("{}/{storage_name}/{key}", owner.path_prefix())
     }
 
     fn create_object(object: &s3::serde_types::Object) -> Object {
@@ -138,9 +148,11 @@ impl StorageClientImpl {
         }
     }
 
-    async fn add_storage(&self, stack_id: StackID, name: &str) -> Result<()> {
-        let path = format!("{METADATA_PREFIX}/{stack_id}/{name}");
-        self.bucket.put_object_stream(&mut &b""[..], path).await?;
+    async fn add_storage(&self, owner: Owner, name: &str) -> Result<()> {
+        if let Owner::Stack(_) = owner {
+            let path = format!("{METADATA_PREFIX}/{}/{name}", owner.path_prefix());
+            self.bucket.put_object_stream(&mut &b""[..], path).await?;
+        }
         Ok(())
     }
 }
@@ -149,25 +161,25 @@ impl StorageClientImpl {
 impl StorageClient for StorageClientImpl {
     async fn update_stack_storages(
         &self,
-        stack_id: StackID,
+        owner: Owner,
         storage_delete_pairs: Vec<(&str, DeleteStorage)>,
     ) -> Result<()> {
-        let existing_storages = self.storage_list(stack_id).await?;
+        let existing_storages = self.storage_list(owner).await?;
 
         for (storage_name, is_delete) in storage_delete_pairs {
             let storage_name = storage_name.to_string();
             if !existing_storages.contains(&storage_name) && !*is_delete {
-                self.add_storage(stack_id, &storage_name).await?;
+                self.add_storage(owner, &storage_name).await?;
             } else if existing_storages.contains(&storage_name) && *is_delete {
-                self.remove_storage(stack_id, &storage_name).await?;
+                self.remove_storage(owner, &storage_name).await?;
             }
         }
 
         Ok(())
     }
 
-    async fn storage_list(&self, stack_id: StackID) -> Result<Vec<String>> {
-        let prefix = format!("{METADATA_PREFIX}/{stack_id}/");
+    async fn storage_list(&self, owner: Owner) -> Result<Vec<String>> {
+        let prefix = format!("{METADATA_PREFIX}/{}/", owner.path_prefix());
 
         let resp = self.bucket.list(prefix, None).await?;
 
@@ -180,27 +192,32 @@ impl StorageClient for StorageClientImpl {
         Ok(objects)
     }
 
-    async fn contains_storage(&self, stack_id: StackID, storage_name: &str) -> Result<bool> {
-        Ok(self
-            .storage_list(stack_id)
-            .await?
-            .contains(&storage_name.into()))
+    async fn contains_storage(&self, owner: Owner, storage_name: &str) -> Result<bool> {
+        match owner {
+            Owner::User(_) => Ok(true),
+            _ => Ok(self
+                .storage_list(owner)
+                .await?
+                .contains(&storage_name.into())),
+        }
     }
 
-    async fn remove_storage(&self, stack_id: StackID, storage_name: &str) -> Result<()> {
+    async fn remove_storage(&self, owner: Owner, storage_name: &str) -> Result<()> {
         // remove from manifest
-        let path = format!("{METADATA_PREFIX}/{stack_id}/{storage_name}");
-        self.bucket.delete_object(path).await?;
+        if let Owner::Stack(_) = owner {
+            let path = format!("{METADATA_PREFIX}/{}/{storage_name}", owner.path_prefix());
+            self.bucket.delete_object(path).await?;
+        }
 
         // remove data
         let keys = self
-            .list(stack_id, storage_name, "")
+            .list(owner, storage_name, "")
             .await?
             .into_iter()
             .map(|o| o.key);
 
         for key in keys {
-            let path = Self::create_path(stack_id, storage_name, &key);
+            let path = Self::create_path(owner, storage_name, &key);
             self.bucket.delete_object(path).await?;
         }
 
@@ -209,49 +226,57 @@ impl StorageClient for StorageClientImpl {
 
     async fn get(
         &self,
-        stack_id: StackID,
+        owner: Owner,
         storage_name: &str,
         key: &str,
         writer: &mut (dyn AsyncWrite + Send + Sync + Unpin),
     ) -> Result<()> {
+        if !self.contains_storage(owner, storage_name).await? {
+            bail!("Storage not found")
+        }
+
         let mut wrapper = AsyncWriterWrapper { writer };
-        let path = Self::create_path(stack_id, storage_name, key);
+        let path = Self::create_path(owner, storage_name, key);
         self.bucket.get_object_stream(path, &mut wrapper).await?;
         Ok(())
     }
 
     async fn put(
         &self,
-        stack_id: StackID,
+        owner: Owner,
         storage_name: &str,
         key: &str,
         reader: &mut (dyn AsyncRead + Send + Sync + Unpin),
     ) -> Result<()> {
-        if !self.contains_storage(stack_id, storage_name).await? {
-            self.add_storage(stack_id, storage_name).await?
+        if !self.contains_storage(owner, storage_name).await? {
+            bail!("Storage not found")
         }
+
         let mut wrapper = AsyncReaderWrapper { reader };
-        let path = Self::create_path(stack_id, storage_name, key);
+        let path = Self::create_path(owner, storage_name, key);
 
         self.bucket.put_object_stream(&mut wrapper, path).await?;
         Ok(())
     }
 
-    async fn delete(&self, stack_id: StackID, storage_name: &str, key: &str) -> Result<()> {
-        let path = Self::create_path(stack_id, storage_name, key);
+    async fn delete(&self, owner: Owner, storage_name: &str, key: &str) -> Result<()> {
+        if !self.contains_storage(owner, storage_name).await? {
+            bail!("Storage not found")
+        }
+
+        let path = Self::create_path(owner, storage_name, key);
 
         self.bucket.delete_object(path).await?;
 
         Ok(())
     }
 
-    async fn list(
-        &self,
-        stack_id: StackID,
-        storage_name: &str,
-        prefix: &str,
-    ) -> Result<Vec<Object>> {
-        let prefix = Self::create_path(stack_id, storage_name, prefix);
+    async fn list(&self, owner: Owner, storage_name: &str, prefix: &str) -> Result<Vec<Object>> {
+        if !self.contains_storage(owner, storage_name).await? {
+            bail!("Storage not found")
+        }
+
+        let prefix = Self::create_path(owner, storage_name, prefix);
 
         let resp = self.bucket.list(prefix, None).await?;
 
@@ -341,7 +366,7 @@ mod test {
 
     use super::*;
 
-    const STACK_ID: StackID = StackID::SolanaPublicKey([1; 32]);
+    const OWNER: Owner = Owner::Stack(StackID::SolanaPublicKey([1; 32]));
 
     async fn test_start() -> Result<Box<dyn StorageManager>> {
         let storage_info = StorageInfo {
@@ -377,11 +402,11 @@ mod test {
             .collect::<Vec<_>>();
 
         client
-            .update_stack_storages(STACK_ID, stor_del_pairs)
+            .update_stack_storages(OWNER, stor_del_pairs)
             .await
             .unwrap();
 
-        let x = client.storage_list(STACK_ID).await.unwrap();
+        let x = client.storage_list(OWNER).await.unwrap();
 
         assert_eq!(insertion_storages, x);
     }
